Free the library!
Modifies how Studio handles libraries so that any advanced problem block
can, at least in principle, be added.  (This was only tested with the
hastexo XBlock.)

The bulk of the changes introduced here are to allow xblocks with
children to be created and edited from within containers displayed in a
library.