Updates to 'free the library'
Update our 'free the library' patch to accomodate some changes made
to the platform between Ironwood and Juniper.
Define variables that are now needed for the container template.
Add a condition for an added XBlockView to be shown only in unit
pages.
Don't allow 'Access Settings' and 'Move' buttons to be added in
library views.
Not all library pages have category set to 'library',
differentiate library pages in container.js from id instead.